Move navbar to components.
